+++ app/export/exportapp.tsx
@@ -1,18 +1,18 @@
 "use client";
 
-import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
-import { FaExternalLinkSquareAlt } from "react-icons/fa";
-import { BeatLoader } from "react-spinners";
+import { useRouter } from 'next/navigation';
+import { useEffect, useState } from 'react';
+import { FaExternalLinkSquareAlt } from 'react-icons/fa';
+import { BeatLoader } from 'react-spinners';
 
-import getPlaylist from "@/actions/spotify/getplaylist";
-import TrackList from "@/components/tracklist";
-import { Button } from "@/components/ui/button";
-import exportCSV from "@/lib/exportcsv";
-import exportJSON from "@/lib/exportjson";
-import exportSimplerCSV from "@/lib/simplerexportcsv";
-import exportSimplerJSON from "@/lib/simplerexportjson";
-import { Playlist, Track } from "@/types";
+import getPlaylist from '@/actions/spotify/getplaylist';
+import TrackList from '@/components/tracklist';
+import { Button } from '@/components/ui/button';
+import exportCSV from '@/lib/exportcsv';
+import exportJSON from '@/lib/exportjson';
+import exportSimplerCSV from '@/lib/simplerexportcsv';
+import exportSimplerJSON from '@/lib/simplerexportjson';
+import { Playlist, Track } from '@/types';
 
 interface ExportAppProps {
 	playlistId?: string;
@@ -111,7 +111,7 @@ const ExportApp = ({ playlistId }: ExportAppProps) => {
 					<h3 className="font-light text-2xl">{header.owner.display_name}</h3>
 					<h1 className="text-7xl font-bold">{header.name}</h1>
 					{header.description && (
-						<h2 className="font-light">{`'${header.description}'`}</h2>
+						<h2 className="font-light">{header.description}</h2>
 					)}
 				</div>
 			</div>